rag-chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai>=0.4.1
qdrant-client==1.9.1
python-dotenv==1.0.0
PyYAML==6.0.1